로그인, 회원가입 이외 페이지 비회원 접근 못하도록 함

+++ MemberAndBoard/src/main/java/handler/loginHandler/DeleteProHandler.java
@@ -14,26 +14,36 @@ public class DeleteProHandler implements CommandHandler {
 		
 		HttpSession session = request.getSession();
 		String id = (String)session.getAttribute("id");
-		String passwd = request.getParameter("passwd");
-		MabDBBean dao = MabDBBean.getInstance();
+		String passwd = (String)session.getAttribute("passwd");
 		String url = null;
-		int result = dao.deleteMember(id, passwd);
+		MabDBBean dao = MabDBBean.getInstance();
+		int checkId = dao.checkLogin(id, passwd);
+		if(checkId == 1) {
 		
-		if(result == 1) {
-			request.setAttribute("result", result);
-			session.setAttribute("id", null);
-			session.setAttribute("passwd", null);
-			url = "/member/deletePro.jsp";
+			passwd = request.getParameter("passwd");
 			
-		}
-		else if(result == -4){
-			request.setAttribute("result", result);
-			url = "/member/errorPage.jsp";
-		}
-		else if(result == 0) {
-			request.setAttribute("result", result);
-			url = "/member/errorPage.jsp";
 			
+			int result = dao.deleteMember(id, passwd);
+			if(result == 1) {
+				request.setAttribute("result", result);
+				session.setAttribute("id", null);
+				session.setAttribute("passwd", null);
+				url = "/member/deletePro.jsp";
+				
+			}
+			else if(result == -4){
+				request.setAttribute("result", result);
+				url = "/member/errorPage.jsp";
+			}
+			else if(result == 0) {
+				request.setAttribute("result", result);
+				url = "/member/errorPage.jsp";
+				
+			}
+		}
+		else {
+			request.setAttribute("checkId", checkId);
+			url = "/member/errorPage.jsp";
 		}
 		
 		return url;
